requirements.txt
python-gammu
sendgrid==6.5.0
pydantic
redis
fastapi
uvicorn